# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Joy VoyagerApi module adds REST Api end points to Voyager with Passport and Swagger support https://github.com/rxcod9/joy-voyager-api.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://finishuseruz.demo",
            "description": "Primary API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Another API Server"
        }
    ],
    "paths": {
...
        "/api/custom-field-by-category/{category}": {
            "get": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Get by category ID",
                "operationId": "e956115f36720800b5ead04cf0f4daaf",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/custom-field-values-by-custom-field/{custom_field}": {
            "get": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Get custom-field by ID",
                "operationId": "360e8ff096242eae7a10a041ccda91cf",
                "parameters": [
                    {
                        "name": "custom_field",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/custom-field-values-by-task/{task}": {
            "get": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Get by task ID",
                "operationId": "96cd8e7408acc1e566878f63b37eff19",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/performers/{performer}": {
            "get": {
                "tags": [
                    "Performers"
                ],
                "summary": "Get list of Performers",
                "operationId": "be33fb1d00766c364214bb291f26b42d",
                "parameters": [
                    {
                        "name": "performer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}